corretto esercizio

+++ app/Http/Controllers/SandwichController.php
@@ -3,21 +3,26 @@
 namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\SandwichRequest;
 use App\Http\Requests\SandwichEditRequest;
+use App\Http\Requests\SandwichRequest;
 use App\Models\Sandwich;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Auth;
 
 class SandwichController extends Controller implements HasMiddleware
 {        
 
-        public static function middleware(): array{
+        public static function middleware(): array
+        {
+
                 return [
-                        (new Middleware('auth'))->only(['create', 'index']),
+
+                        new Middleware('auth', except: ['index']),
+
                 ];
-        }  
+                
+        }
 
         public function index(){
                 $sandwiches = Sandwich::all();
@@ -57,32 +62,62 @@ class SandwichController extends Controller implements HasMiddleware
 
         public function edit(Sandwich $sandwich){
 
-                return view('sandwich.edit', compact('sandwich'));
+                if ($sandwich->user_id == Auth::user()->id) {
+
+                        return view('sandwich.edit', compact('sandwich'));   
+
+                }else{
+
+                        return redirect()->route('home')->with('errorMessage', 'Non puoi accedere a questa pagina');
+                        
+                }  
         
         }
 
         public function update(SandwichEditRequest $request, Sandwich $sandwich){
 
-                $sandwich->update([
-                        $sandwich->name = $request->name,
-                        $sandwich->description = $request->description,
-                ]);
+                if ($sandwich->user_id == Auth::user()->id) {
 
-                if($request->img){
                         $sandwich->update([
-                                $sandwich->img = $request->file('img')->store('public/images')
-                        ]);
-                }
 
-                return redirect()->route('home')->with('successMessage', 'Hai modificato correttamente il tuo panino');
+                                $sandwich->name = $request->name,
+                                $sandwich->description = $request->description,
+
+                        ]);
+
+                        if($request->img){
+
+                                $sandwich->update([
+
+                                        $sandwich->img = $request->file('img')->store('images', 'public')
+
+                                ]);
+                        }
+
+                        return redirect()->route('home')->with('successMessage', 'Hai modificato correttamente il tuo panino');   
+
+                }else{
+
+                        return redirect()->route('home')->with('errorMessage', 'Non puoi accedere a questa pagina');
+
+                }
+                
 
         }
 
         public function destroy(Sandwich $sandwich){
 
-                $sandwich->delete();
+                if ($sandwich->user_id == Auth::user()->id) {
 
-                return redirect()->route('home')->with('successMessage', 'Hai eliminato correttamente il tuo panino');
+                        $sandwich->delete();
+
+                        return redirect()->route('home')->with('successMessage', 'Hai eliminato correttamente il tuo panino');   
+
+                }else{
+
+                        return redirect()->route('home')->with('errorMessage', 'Non puoi accedere a questa pagina');
+
+                }            
 
         }
